backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.7.1
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
jinja2>=3.1.4
httpx>=0.27.0
